Move create_config to Config::create_config
